backend/requirements.txt
fastapi 
uvicorn
groq
python-dotenv
redis
sentence-transformers
numpy
langchain
langchain-groq
langchain-google-genai



